Add files via upload
There are 5 different files having different classes for performing different tasks. It is developed using python 3.6 and its library tkinter.
1-class_main
it is makes a window appear on the screen asks for login initials of users and if they are present in the database it sends user to the respective window on the basis of role present in row where password and username are present. User doesn’t have to mention its role.
2-class_fac
It displays a frame with 3 buttons MCQs, T/F, Numeric. Teacher can select desired type and has to write just question and answer in T/F and Numeric and for MCQs teacher can add up to 4 options.
3-class_stud
It displays all the questions and options for student and buttons as option a, b, c, d and checks answer on the basis of option clicked and adds 0 or 1 in the score on the basis of result and also informs this question is attempted. If a question is attempted again it overwrites previous answer. Save button leads to new frame that shows total score.
4-class_show
It simply displays total number of all the question attempted correctly
5-class_unittest
It has some unit tests to check if same number of scores are available as number of students, same number of questions are present as answers, and number of roles, passwords and usernames is equal.